Repackage after KDE Frameworks 5.71 update
This package has some optional dependencies, and if those dependencies
are not present in the build environment, or do not satisfy declared
version constraints, then a plugin (kf5/kirigami/org.kde.desktop.so)
will not be built.

If KDE packages are updated in dependency order, then packaging problems
of this kind are less likely to occur. But for now, just re-autospec to
ship the plugin.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>